requirements_dev.txt
pip==22.0.2
numpy==1.24.3
trimesh==4.6.13
pytest==6.2.4
streamlit-image-coordinates==0.2.2
Pillow==10.0.1
matplotlib==3.7.2
streamlit-stl==0.0.6
scipy==1.10.1
shapely==2.0.7
pyvista==0.44.2
imageio==2.35.1
networkx
mapbox-earcut
